The HI to H/C conversion is put back into the fastgenex6 code along with retaining the old function for this conversion that is an eight parameter polynomial.
#253502
@<Erik.vandenBoogaard> I have attached the conversion function formula and coefficient values used in the code in the related PBI for your benefit

Related work items: #253502

+++ development/libraries/casaAPI/src/RSProxyQualityCalculator.cpp
@@ -186,7 +186,7 @@ std::vector<double> RSProxyQualityCalculator::calculateQ2(const string& proxyNam
 
 const std::vector<std::vector<double>> RSProxyQualityCalculator::calculateR2AndR2adjFromObservables(const std::vector<std::vector<double>>& runCaseObservables, const std::vector<std::vector<double>>& proxyObservables) const
 {
-  const int nTargets{runCaseObservables[0].size()};
+  const int nTargets{ static_cast<int>(runCaseObservables[0].size())};
 
   std::vector<std::vector<double>> R2AndR2adj(2, std::vector<double>(nTargets, 0.0));
   std::vector<double> yAvg = averagesOfMatrixColumns(runCaseObservables);
